requirements.txt
aiohttp
asyncio
discord.py
sendgrid==6.4.5
requests==2.25.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0